requirements/dev.txt
-r base.txt

django-debug-toolbar==1.6
sqlparse==0.2.1
